change flet again and refactor login page
